Update: Allow custom positioning for anchored tags
With this commit, the position of anchored-tags can be customized
i.e. anchored tags can appear *after* (the default) or *before*
headings.

This commit adds a parameter -- `Params.AnchorTagPositionLeft`.
This is a boolean indicating if the anchor tag is to be positioned
on the left of the title. Using `false` will set the anchor tag to
be at the right.

The default behaviour is the same as before - anchor tags will
appear *after* headings (to the right), unless moved to the left
using this new parameter