Update: Add a `future` tag for future posts
This commit adds a `future` tag as a visual indicator for posts
that are scheduled for a release in the future. In terms of looks,
this tag is the same as the existing `draft` tag for posts marked
as draft.

To decide if a post is to be released in the future, the template
will compare `Page.Date` and `Page.PublishDate` fields from the
front matter, checking if either one is greater than the current
time. If yes, the post will be tagged as a `future` post