Minor: Small tweaks. No visible changes
The previous commit elevated the entire `entry-tags-container`,
ensuring that the container would not have links - i.e. tapping
on the container will have no effect.

The container itself expanded to fill the width of the article
container. Meaning that there would be a thin strip of empty area
inside the article block that would not contain any hyperlink

This commit fixes that up. Instead of elevating the entire
container and making it lose the hyperlink, this commit ensures
that the tags are selectively being elevated. Meaning there is
no longer a strip of area within an article block that would not
contain a link.

This is just a minor UX-fix, and is not supposed to affect the
existing functionality of the theme