Feature: Define custom popup for `abbr` tag
This commit adds a custom popup that can be used as a replacement
for the default `abbr` tag. The modifications made include better
UI for the popup tooltip generated, and support for mobile/touch
devices

Using this modified version is as simple as
	<abbr full-form="World Wide Web">www</abbr>

And, if needed, replacing this `abbr` tag with the default tag is
as simple as
	<abbr title="World Wide Web">www</abbr>

The only change being the `title`/`full-form` fields!

Source;
https://bitsofco.de/making-abbr-work-for-touchscreen-keyboard-mouse