Update: Display an additional listing page as home page
The main difference between the "home" page, and any usual lisiting
page were the missing site title, summary and social icon widgets,
or simply — the `home_info.html` partial

This commit intorduces an internal variable `isCustomHome` that
checks the path to know if a page is supposed to be used as a
custom home page or not. If yes, the missing elements - i.e. the
`home_info` partial is also displayed on the page, making the page
be an additional home page

Signed-off-by: notsatan <git@notsatan.me>